second major update beat tracker

+++ src/audio.h
@@ -8,8 +8,6 @@
 #include <cstring>
 #include <unistd.h>
 
-// int streamAudio (jack_nframes_t nframes, void *arg);
-
 class AudioStream{
 public:
 
@@ -24,15 +22,13 @@ public:
 
     bool trigger_chunk;
 
-    // int get_signal(float *in,void (*)(float*));
-
 private:
     jack_port_t *input_port;
     jack_port_t *output_port_left;
     jack_port_t *output_port_right;
     jack_client_t *client;
 
-    float *in;
+    double *in;
     
     const char *server_name;
     const char *client_name ;
@@ -40,8 +36,7 @@ private:
     const char **todevice;
 
     static void jack_shutdown (void *arg);
-    //init BTrack
-    // int streamAudio(jack_nframes_t nframes, void *arg); //float *buffer 
+
 };
